Reorganization *read desc*
Optimized the structure of the program. Many file management operations are now in util.py as well as a general decluttering of most functions
Comments galore!
updater.py now just stores the questions to ask the user and stuff like that
updated requirements.txt
Moved UpdateCode to vars.py, and added a GAME_PATH const
Added traceback printing to the print_exception_error_dbg func
Updated the UpdateInfo things to accept a hotfix flag so that it remembers if it needs to update from the 0.7-hotfix verison